// src/app/shared/components/comment-section/comment.interface.ts
// "id": "276763566129324f90e525140f3908ce1b94cc4446d21455",
// "question_id": "10896454193-0530-447a-abbb-fc09306541b7",
// "comment": "LO",
// "comment_by": "27676356612",
// "comment_at": "2024-02-18T22:15:42.353518",
// "created_at": "2024-02-18T20:15:43.161951",
// "updated_at": "2024-02-18T20:15:43.161951"

export interface Comment {
    id?: string;
    question_id?: string;
    comment?: string;
    comment_by?: string;
    created_at?: string;
    updated_at?: string;
    success?: string;
    phone?: string;
    user_id?: number;
    color?: string;
}